# uStoreAPI/uStoreAPI/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "uStoreAPI, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Tiendas/GetTiendas": {
      "get": {
        "tags": [
          "Tiendas"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "idCentroComercial",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/TiendaDto"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Tiendas/GetImagenesTienda": {
      "get": {
        "tags": [
          "Tiendas"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "idTienda",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/TiendaDto"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Tiendas": {
      "get": {
        "tags": [
          "Tiendas"
        ],
        "operationId": "GetTienda",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/TiendaDto"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Tiendas/CreateTienda": {
      "post": {
        "tags": [
          "Tiendas"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "IdTienda",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "NombreTienda",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "RangoPrecio",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "IdCentroComercial",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "IdAdministrador",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "Apartados",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "Vistas",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "LogoTienda",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "logoTienda",
            "type": "file"
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/TiendaDto"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Tiendas/CreateImagenTienda": {
      "post": {
        "tags": [
          "Tiendas"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "idTienda",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          },
          {
            "in": "formData",
            "name": "imagen",
            "type": "file"
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/ImagenesTienda"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Tiendas/UpdateImagenTienda": {
      "put": {
        "tags": [
          "Tiendas"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "idTienda",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "idImagenTienda",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          },
          {
            "in": "formData",
            "name": "imagen",
            "type": "file"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Tiendas/UpdateTienda": {
      "put": {
        "tags": [
          "Tiendas"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "IdTienda",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "NombreTienda",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "logoTienda",
            "type": "file"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Tiendas/DeleteTienda": {
      "delete": {
        "tags": [
          "Tiendas"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Tiendas/DeleteImagenTienda": {
      "delete": {
        "tags": [
          "Tiendas"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CalificacionTiendum": {
      "type": "object",
      "properties": {
        "idCalificacionTienda": {
          "format": "int32",
          "type": "integer"
        },
        "calificacion": {
          "format": "int32",
          "type": "integer"
        },
        "idTienda": {
          "format": "int32",
          "type": "integer"
        },
        "idUsuario": {
          "format": "int32",
          "type": "integer"
        },
        "idTiendaNavigation": {
          "$ref": "#/definitions/Tiendum"
        },
        "idUsuarioNavigation": {
          "$ref": "#/definitions/Usuario"
        }
      },
      "additionalProperties": false
    },
...
    "CategoriasTienda": {
      "type": "object",
      "properties": {
        "idCt": {
          "format": "int32",
          "type": "integer"
        },
        "idTienda": {
          "format": "int32",
          "type": "integer"
        },
        "idCategoria": {
          "format": "int32",
          "type": "integer"
        },
        "idCategoriaNavigation": {
          "$ref": "#/definitions/Categoria"
        },
        "idTiendaNavigation": {
          "$ref": "#/definitions/Tiendum"
        }
      },
      "additionalProperties": false
    },
...
    "Comentario": {
      "type": "object",
      "properties": {
        "idComentarios": {
          "format": "int32",
          "type": "integer"
        },
        "comentario1": {
          "type": "string"
        },
        "idProductos": {
          "format": "int32",
          "type": "integer"
        },
        "idTienda": {
          "format": "int32",
          "type": "integer"
        },
        "idUsuario": {
          "format": "int32",
          "type": "integer"
        },
        "idProductosNavigation": {
          "$ref": "#/definitions/Producto"
        },
        "idTiendaNavigation": {
          "$ref": "#/definitions/Tiendum"
        },
        "idUsuarioNavigation": {
          "$ref": "#/definitions/Usuario"
        }
      },
      "additionalProperties": false
    },
...
    "Favorito": {
      "type": "object",
      "properties": {
        "idFavoritos": {
          "format": "int32",
          "type": "integer"
        },
        "idProductos": {
          "format": "int32",
          "type": "integer"
        },
        "idTienda": {
          "format": "int32",
          "type": "integer"
        },
        "idUsuario": {
          "format": "int32",
          "type": "integer"
        },
        "idProductosNavigation": {
          "$ref": "#/definitions/Producto"
        },
        "idTiendaNavigation": {
          "$ref": "#/definitions/Tiendum"
        },
        "idUsuarioNavigation": {
          "$ref": "#/definitions/Usuario"
        }
      },
      "additionalProperties": false
    },
...
    "ImagenesTienda": {
      "type": "object",
      "properties": {
        "idImagenesTiendas": {
          "format": "int32",
          "type": "integer"
        },
        "imagenTienda": {
          "type": "string"
        },
        "idTienda": {
          "format": "int32",
          "type": "integer"
        },
        "idTiendaNavigation": {
          "$ref": "#/definitions/Tiendum"
        }
      },
      "additionalProperties": false
    },
...
    "ProblemDetails": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        }
      },
      "additionalProperties": { }
    },
...
    "TiendaDto": {
      "required": [
        "idCentroComercial",
        "nombreTienda"
      ],
      "type": "object",
      "properties": {
        "idTienda": {
          "format": "int32",
          "type": "integer"
        },
        "nombreTienda": {
          "minLength": 1,
          "type": "string"
        },
        "rangoPrecio": {
          "type": "string"
        },
        "idCentroComercial": {
          "format": "int32",
          "type": "integer"
        },
        "idAdministrador": {
          "format": "int32",
          "type": "integer"
        },
        "apartados": {
          "format": "int32",
          "type": "integer"
        },
        "vistas": {
          "format": "int32",
          "type": "integer"
        },
        "logoTienda": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "Tiendum": {
      "type": "object",
      "properties": {
        "idTienda": {
          "format": "int32",
          "type": "integer"
        },
        "nombreTienda": {
          "type": "string"
        },
        "rangoPrecio": {
          "type": "string"
        },
        "idCentroComercial": {
          "format": "int32",
          "type": "integer"
        },
        "idAdministrador": {
          "format": "int32",
          "type": "integer"
        },
        "apartados": {
          "format": "int32",
          "type": "integer"
        },
        "vistas": {
          "format": "int32",
          "type": "integer"
        },
        "logoTienda": {
          "type": "string"
        },
        "calificacionTienda": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CalificacionTiendum"
          }
        },
        "categoriasTienda": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CategoriasTienda"
          }
        },
        "comentarios": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Comentario"
          }
        },
        "favoritos": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Favorito"
          }
        },
        "gerentes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Gerente"
          }
        },
        "historials": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Historial"
          }
        },
        "horarios": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Horario"
          }
        },
        "idAdministradorNavigation": {
          "$ref": "#/definitions/AdministradorTiendum"
        },
        "idCentroComercialNavigation": {
          "$ref": "#/definitions/CentroComercial"
        },
        "imagenesTienda": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ImagenesTienda"
          }
        },
        "periodosPredeterminados": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/PeriodosPredeterminado"
          }
        },
        "productos": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Producto"
          }
        },
        "publicaciones": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Publicacione"
          }
        },
        "solicitudesApartados": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SolicitudesApartado"
          }
        },
        "tendenciasVenta": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TendenciasVentum"
          }
        }
      },
      "additionalProperties": false
    },
...
  }
}